requirements.txt
# requests
# requests-cache
# wikitextparser
pyyaml